posts/requirements.txt
apt-clone==0.2.1
apturl==0.5.2
Babel==2.3.4
beautifulsoup4==4.7.1
braintree==3.52.0
Brlapi==0.6.6
catfish==1.4.4
certifi==2019.3.9
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
cupshelpers==1.0
decorator==4.0.10
defer==1.0.6
Django==2.2
djangorestframework==3.9.4
docutils==0.12
ebaysdk==2.1.5
fail2ban==0.10.2
feedparser==5.2.1
gevent==1.1.2
greenlet==0.4.10
html2text==2016.9.19
httplib2==0.9.2
idna==2.7
Jinja2==2.10.1
libsass==0.12.3
louis==3.5.0
lxml==3.7.1
Mako==1.0.4
MarkupSafe==0.23
menulibre==2.2.0
mock==2.0.0
mugshot==0.4.0
netifaces==0.10.4
num2words==0.5.6
numpy==1.16.4
odoo==12.0.post20190508
ofxparse==0.16
olefile==0.46
onboard==1.4.1
pandas==0.24.2
passlib==1.6.5
pbr==5.2.0
pexpect==4.2.1
Pillow==4.0.0
psutil==4.3.1
psycopg2==2.7.3.1
psycopg2-binary==2.8.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pydot==1.2.3
pygobject==3.26.1
PyICU==1.9.8
pyldap==2.4.28
pyparsing==2.1.10
PyPDF2==1.26.0
pyserial==3.1.1
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.32
python-stdnum==1.11
python-xapp==1.4.0
pytz==2019.1
pyusb==1.0.0
pyxdg==0.25
PyYAML==3.12
qrcode==5.3
reportlab==3.3.0
requests==2.20.0
requests-unixsocket==0.1.5
roman==2.0.0
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
suds-jurko==0.6
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.24.3
vatnumber==1.2
vobject==0.9.3
Werkzeug==0.11.15
xkit==0.0.0
xlrd==1.0.0
XlsxWriter==0.9.3
xlwt==1.3.0
